apps/api/pyproject.toml
[project]
name = "manager-jarvis-api"
version = "0.1.0"
description = "Manager Jarvis Docker management API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "sqlalchemy>=2.0.36",
  "pydantic-settings>=2.7.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "docker>=7.1.0",
  "python-multipart>=0.0.18",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "httpx>=0.27.2",
  "ruff>=0.8.0",
  "mypy>=1.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
